move URI SA type to deprecated namespace
