+++ db/functions/packed.py
@@ -7,8 +7,7 @@ Mathesar filters not supporting composition.
 from abc import abstractmethod
 
 from db.functions import base
-from db.types.custom.uri import URIFunction
-from db.deprecated.types.custom import EMAIL_DOMAIN_NAME
+from db.deprecated.types.custom import EMAIL_DOMAIN_NAME, URIFunction
 
 
 class DBFunctionPacked(base.DBFunction):
